os_tools: execute: cleanup and mimic old behaviour on failure
oe.execute() returned an empty string on failure if a result was wanted.

For the exit code, it's the shell's exit code instead of the command's,
so it may be misleading if not reading closely. The command's error
output will be more useful, and remains.

For the FileNotFoundError exception, this is whether or not the shell
is found. Bigger problems are afoot if python can't find a shell.

Signed-off-by: Ian Leonard <antonlacon@gmail.com>