updates: rework rpi firmware update
The RPi firmware update buttons are still an true/false toggle. It is
set to true if the user initiated an update, tries to toggle it off
after initiating an update (already queued), or restarts Kodi (but not
the machine) with an update queued.

It's false if the user selects no, the firmware is current, or no update
is queued.

If a user wishes to cancel an update after it's been queued, they need to
ssh in and enter 'rpi-eeprom-update -r'. User would need to restart Kodi
or select the update and choose 'no' to get buttons into matching state
again.

Signed-off-by: Ian Leonard <antonlacon@gmail.com>